requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
openai
python-dotenv
PyYAML